pyproject.toml
[project]
name = "parakeet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "onnx-asr[cpu,hub]>=0.7.0",
    "pyaudio>=0.2.14",
    "pydbus>=0.6.0",
    "pygobject>=3.54.3",
    "pillow>=10.0.0",
    "cairosvg>=2.7.0",
]
